+++ CHANGES.rst
@@ -5,7 +5,7 @@ CHANGES
 ------
 - Fixed a problem with django-logdb test suite being used in a project with
   an existing logging configuration.
-
+- Updated testproject to Django 1.2.
 
 0.9.9
 -----